feat: make reader.py, writer.py, and processor.py into stand-alone python programs.
Do this to enable each of these items to be run by a Galaxy tool in the future.
NB: the `--data` option used for Processors and Writers still presumes to simply
parse & pass a string. This is fine for the toy model, but will likely be
insufficient for the full implementation. Address it later.