fix: reference data shown when selecting material parameteres in edd Processors
Use the sum of all MCA spectra as the reference plot intead of just
the 0th spectrum. Selecting just one spectrum for reference is
precarious because the sample may not have been in the beam (in which
case: diffraction peaks are not visible). Sum of all spectra is more
likely to be a useful reference plot for sleecting material
parameters.