fix: override SMBMCAScanParser.get_spec_scan_motor_vals
Idiosyncraticies in EDD data collection make it impossible (currently)
to consistently get correct values for scanned motor values from the
data available in the _spec file_. However, it is possible to
consistently get the correct _relative_ motor positions, but only from
data available in the scan's _.par file_, and with a meaning of
"relative" that is specific to EDD experiments. Since the relative
motor values are the only ones currently used in the EDD workflow,
make those the ones available in
`SMBMCAScanParser.get_spec_scan_motor_vals`.