refactor: Splie the Ceria calib configs in energy and tth configs
Move the parameters peak_energies, max_peak_index and fit_index_ranges
to the energy pydantic config.
Allow tth_initial_guess and include_energy_ranges as inputs to
edd.MCATthCalibrationProcessor to override the values of the
corresponding fields in the energy calib input yaml file for all
detectors.
Removed hkl_indices from the input config. Made it a private attr for
now, but it's not really needed as part of the configuration.